requirements.txt
absl-py==0.13.0
astunparse==1.6.3
bert-for-tf2==0.14.9
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cycler==0.10.0
ekphrasis==0.5.1
emoji==1.4.2
flatbuffers==1.12
ftfy==6.0.3
gast==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.2
importlib-metadata==4.6.3
joblib==1.0.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
nltk==3.6.2
numpy==1.21.1
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.0
params-flow==0.8.2
Pillow==8.3.1
protobuf==3.17.3
py-params==0.10.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
sklearn==0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
threadpoolctl==2.2.0
tqdm==4.62.0
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
